llm/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "llm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=1.93.0",
    "streamlit>=1.46.1",
    "python-dotenv>=1.0.0",
]

[project.scripts]
dev = "streamlit:main"
